+++ resolwe_bio/processes/microarray/methylation_arrays.py
@@ -2,8 +2,6 @@
 from pathlib import Path
 from shutil import copy2
 
-import pandas as pd
-from numpy import nan
 from plumbum import TEE
 
 from resolwe.process import (
@@ -36,7 +34,7 @@ class MethylationArraySesame(ProcessBio):
     slug = "methylation-array-sesame"
     name = "Methylation analysis (SeSAMe)"
     process_type = "data:methylation:sesame"
-    version = "1.1.0"
+    version = "1.2.0"
     category = "Methylation arrays"
     data_name = 'SeSAMe array ({{ idat_file.red_channel.file|default("?") }})'
     scheduling_class = SchedulingClass.BATCH
@@ -98,56 +96,7 @@ class MethylationArraySesame(ProcessBio):
                 "An error was encountered during the running of SeSAMe pipeline."
             )
 
-        meth_data = "beta_values_annotated.txt.gz"
-        xy = pd.read_csv(
-            filepath_or_buffer=meth_data,
-            sep="\t",
-        )
-
-        mapping_filters = {"source_db": "ILLU_METHYL", "target_db": "ENSEMBL"}
-        mappings = self.mapping.filter(**mapping_filters)
-        ens = [
-            {
-                "source_id": mapping.source_id,
-                "target_id": mapping.target_id,
-            }
-            for mapping in mappings
-        ]
-
-        ens = pd.DataFrame(ens)
-
-        mapped_ids = []
-        for multigene in xy.gene_HGNC:
-            if multigene is nan:
-                mapped_ids.append(nan)
-                continue
-
-            multi_split = multigene.split(";")
-            mapped_tmp = []
-            for gene in multi_split:
-                ids = ens[ens.source_id == gene].target_id
-                if len(ids) == 0:
-                    mapped_tmp.append("")
-                    continue
-                mapped_tmp.append(",".join(ids))
-            # In case there are two genes with no mappings in KB, the
-            # result will look like " , " (missing values left and
-            # right). If there is only one gene (no splitting by a ,),
-            # an empty value is present.
-            mapped_tmp = ",".join(mapped_tmp)
-            mapped_ids.append(mapped_tmp)
-
-        xy["ensembl_id"] = mapped_ids
-
-        xy = xy[["probe_ids", "gene_HGNC", "betas", "mvals", "pvals", "ensembl_id"]]
-
-        xy.to_csv(
-            path_or_buf=meth_data,
-            sep="\t",
-            index=False,
-        )
-
         outputs.qc_data = "QC_data.txt"
-        outputs.methylation_data = meth_data
+        outputs.methylation_data = "beta_values_annotated.txt.gz"
         outputs.species = inputs.idat_file.output.species
         outputs.platform = platform